language-python.ts split out export code #2813
Move the bulky code and tables for translating the library function calls in expressions when exporting.
Create three new files:
(1) src/ide/frames/frame-interfaces/language-export.ts new Interface LanguageExport
  implemented by language-export-python.ts (and for other languages in due course perhaps)
  These are the language-specific classes which hold the lookup tables of Elan and other language
function names and parameter order.
(2) src/ide/frames/language-export-helpers.ts new helper class LanguageExportHelpers
which holds the code common to all languages which parses the expressions and translates
them according to the data in the tables.
(3) src/ide/frames/language-export-python.ts new class LanguageExportPython which implements the LanguageExport interface(1) with lookup tables specific to Python.
These files are mostly populated with code extracted (and deleted) from src/ide/frames/language-python.ts.
The entry points `renderFileImportsAsHtml` and `translateExpression` remain in language-python.ts, but they are now very thin and just call straight across to LanguageExportHelpers(2) passing an instance of the Python-specific tables class(3).
The functionality is broadly unchanged.
I have done one or two small fixes:
- add a variable `enableExportTranslation` in language-export-helpers.ts so that the entire functionality (for all languages) can be turned on and off by changing `true` `<->` `false`.
- turn round the brackets on the fly in the regexp in `translateStandaloneWithParams`.
- update the `parseAsInt` translation to just map to `int`.
- re-implement the import handling to separate the language-independent part.  New function `getImportNeeded` in the language-specific part.
- re-implement the processing of `equals` to be table-driven.
- make the `math` prefix configurable per language, it isn't always `math.`.
- I have not touched `renderFileTrailerAsHtml` (the call to `main()`) as it is trivial and not really to do with translating library function calls in expressions. It remains in `src/ide/frames/language-python.ts`.
I've made no attempt to carry the git history over to the new files.
